sanchai-weather/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
langchain
langchain-openai
langchain-core


